+++ R/create_hexbins.R
@@ -8,8 +8,10 @@
 #' @export
 #' @examples
 #' \dontrun{
-#' get_hexbins('LSO', 1)
-#' get_hexbins('UGA', 2)
+#'  library(gisr)
+#'
+#'  get_hexbins('LSO', 1)
+#'  get_hexbins('UGA', 2)
 #' }
 #'
 get_hexbins <- function(country_code,