+++ R/generate_hexbins.R
@@ -6,8 +6,13 @@
 #' @export
 #' @examples
 #' \dontrun{
-#' generate_hexbins(data)
-#' generate_hexbins(data, 10000)
+#'  library(gisr)
+#'
+#'  cntry <- "Nigeria"
+#'  data <- get_admin0(cntry)
+#'
+#'  generate_hexbins(data)
+#'  generate_hexbins(geodata = data, size = 10000)
 #' }
 #'
 generate_hexbins <- function(geodata, size=NULL) {